git: add pre-push hook
This pre-push hook invokes cppcheck only for the files touched by the branch
being pushed with respect to upstream's develop branch. Furthermore, only
files inside ./src/ are checked.

This hook assumes the upstream remote to be called 'upstream'. If not
existent it falls back to the remote provided on the git push command line.

The cppcheck message about missing includes is currently suppressed because
of frequent changes in the repo structure.